refactor: Use a small src/index.ts so that lib/index.js doesn't change often
